// src/interfaces/Estado.type.ts
export interface Estado {
  cdEstado?: number;
  nmEstado?: string;
  cdAgencia?: number;
  nmAgencia?: string;
  cdUnidade?: number;
  nmUnidade?: string;
  cdFormato?: number;
  nmFormato?: string;
  cdClassificacao?: number;
  nmClassificacaoPt?: string;
  nmClassificacaoEn?: string;
  cdNmColuna?: string;
  nmDescricaoPt?: string;
  nmDescricaoEn?: string;
  nmLabelPt?: string;
  nmLabelEn?: string;
  vlPorCd?: string;
}